Fix playback of live broadcasts
Telekom now prevents sharing internal stream urls.
Cookies have to be set to be able to download informations for substreams.

Playback currently only works with serviceApp and exteplayer3.
In gstreamer(like it is used in e2) you can't set user-agent and cookies for
all connections. It falls back to default values and so connection don't
work.